fix(firefox-download): Fix ws disconnecting when downloading
When we download a file, we create a temporary link with a
download attribute. We then programatically click on it,
on Firefox this closes any ongoing WebSocket connection which
causes the application to go in a "disconnected" state.

See here for additional informations:

https://bugzilla.mozilla.org/show_bug.cgi?id=858538
https://bugzilla.mozilla.org/show_bug.cgi?id=896666

Until this is fixed, we added a `target="_blank"` attribute
as a workaround.