ci: move the deploy to a new account and buckets
As discussed over Discord, we need to move the S3 buckets that are
hosting https://app.fluidd.xyz and https://dev-app.fluidd.xyz to a new
account.
I took the opportunity to replace the deprecated tool to upload to S3
with a github action that supports a more modern authentication method.

There are a few secrets to be added to the repository settings which I
will share via private message.

Signed-off-by: Rogerio Goncalves <rogerlz@gmail.com>